+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Notable changes to the `snowdrop-id`, `snowdrop-id-cli`, and
 [Keep a Changelog](https://keepachangelog.com); versions follow
 [SemVer](https://semver.org).
 
-## [Unreleased]
+## [0.3.0] - 2026-07-16
 
 ### Breaking